Break up room mosule
